refactor(codebuddy): rename provider to codebuddy-cn + add logo
codebuddy.cn and codebuddy.ai are distinct deployments (different host,
endpoint, models, auth realm). Rename the existing provider (the .cn /
copilot.tencent.com variant) from `codebuddy` to `codebuddy-cn` across
the registry, OAuth flow, token refresh, executor, and OAuth constants
so a future `codebuddy-ai` provider can coexist. Add the official
CodeBuddy logo at public/providers/codebuddy-cn.png.
