+++ open-sse/providers/registry/index.js
@@ -16,7 +16,7 @@ import p13 from "./chutes.js";
 import p14 from "./claude.js";
 import p15 from "./cline.js";
 import p16 from "./cloudflare-ai.js";
-import p17 from "./codebuddy.js";
+import p17 from "./codebuddy-cn.js";
 import p18 from "./codex.js";
 import p19 from "./cohere.js";
 import p20 from "./comfyui.js";
