+++ tests/__baseline__/alias-baseline.json
@@ -105,7 +105,7 @@
     "claude": "cc",
     "cline": "cl",
     "cloudflare-ai": "cloudflare-ai",
-    "codebuddy": "codebuddy",
+    "codebuddy-cn": "codebuddy-cn",
     "codex": "cx",
     "cohere": "cohere",
     "commandcode": "commandcode",
